More primitive elim form for inductives
Previous eliminator for inductive types was too complex. It performs
automagic currying, making it difficult to type check and reduce.

New version must be fully applied, and magic is left to the surface language.

This version is now working and all core tests are passing, but
documentation, etc needs to be changed. Also, there is dead code that
should be removed.