Split subtyping/reduce in conversion, in model
Previous, conversion was doing both reduction and subtyping.
Now, conversion first reduces, then appeal to the subtyping judgment.
The subtyping judgment may recur on conversion for certain subtyping
rules.
This is a faithful model of the CIC spec.

Unfortunately, after a little bisecting, this seems to introduce a
pretty serious performance bug; 2x running time.
This is not much of an issue since this change is in the faithful model,
not the Redex implementation.