Tweaks
Seems the issue with stlc.rkt is a performance bug, not an infinite
loops. Not sure what about parameters causes this bug, but it can be
made worse by reordering certain recursive judgments...
This is demonstrated by commenting out the type-check line for checking methods.
In the current configuration, this will terminate in about 2 minutes.
After reordering the recursive case of the wf judgment, it does not
terminate after several hours.