Fixed define-datatype and elim-D syntax
Previously, thought define-datatype expected constructors of the form:

  c-name : (Π ([A : AT] ...)
                    (Π ([I : IT] ...)
                       (Π ([r : rT] ...)
                          c-result)))
  where A ... are the parameters, I ... are the indices, and r ... are everything else.

Experimentation with turnstile indicates that this is not the case; it only requires A... and then r
..., where r ... includes indices.
Now dep-data generates the right syntax.

Also updated the syntax for elim-D (turnstile's eliminators). elim-D methods are of the form:

  (λ ([I : IT] ...)
    (λ ([r : rT] ...)
      (λ ([IH : IHT] ...)
         body)))
  where I ... and r ... are as above, and IH ... are the inductive hypothesis generated by the
  recursive arguments.

regardless of the whether the constructor has any indices or recursive arguments.
Updated this in the test, but we need abstractions to hide this from Cur users.